Remove depedency on System.Windows.Interactivity (reverted from commit 67e45787ae33a6d97fb16195db51347cd46218d4)
