+++ RetailCoder.VBE/UI/Controls/BindableSelectedItemBehavior.cs
@@ -1,0 +1,53 @@
+using System.Windows;
+using System.Windows.Controls;
+using System.Windows.Interactivity;
+
+namespace Rubberduck.UI.Controls
+{
+    /// <summary>
+    /// http://stackoverflow.com/a/5118406/1188513
+    /// </summary>
+    public class BindableSelectedItemBehavior : Behavior<TreeView>
+    {
+        public object SelectedItem
+        {
+            get { return (object) GetValue(SelectedItemProperty); }
+            set { SetValue(SelectedItemProperty, value); }
+        }
+
+        public static readonly DependencyProperty SelectedItemProperty =
+            DependencyProperty.Register("SelectedItem", typeof (object), typeof (BindableSelectedItemBehavior),
+                new UIPropertyMetadata(null, OnSelectedItemChanged));
+
+        private static void OnSelectedItemChanged(DependencyObject sender, DependencyPropertyChangedEventArgs e)
+        {
+            var item = e.NewValue as TreeViewItem;
+            if (item != null)
+            {
+                item.SetValue(TreeViewItem.IsSelectedProperty, true);
+            }
+        }
+
+        protected override void OnAttached()
+        {
+            base.OnAttached();
+
+            AssociatedObject.SelectedItemChanged += OnTreeViewSelectedItemChanged;
+        }
+
+        protected override void OnDetaching()
+        {
+            base.OnDetaching();
+
+            if (AssociatedObject != null)
+            {
+                AssociatedObject.SelectedItemChanged -= OnTreeViewSelectedItemChanged;
+            }
+        }
+
+        private void OnTreeViewSelectedItemChanged(object sender, RoutedPropertyChangedEventArgs<object> e)
+        {
+            SelectedItem = e.NewValue;
+        }
+    }
+}
