--- RetailCoder.VBE/UI/Controls/SelectableTreeView.xaml
@@ -1,7 +1,0 @@
-﻿<TreeView x:Class="Rubberduck.UI.Controls.SelectableTreeView"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-             mc:Ignorable="d">
-</TreeView>
