--- RetailCoder.VBE/UI/Controls/SelectableTreeView.xaml.cs
@@ -1,37 +1,0 @@
-﻿using System.Windows;
-using System.Windows.Controls;
-
-namespace Rubberduck.UI.Controls
-{
-    public partial class SelectableTreeView : TreeView
-    {
-        public SelectableTreeView()
-        {
-            InitializeComponent();
-        }
-
-        public new object SelectedItem
-        {
-            get { return (object)GetValue(SelectedItemProperty); }
-            set { SetValue(SelectedItemProperty, value); }
-        }
-
-        public static readonly new DependencyProperty SelectedItemProperty =
-            DependencyProperty.Register("SelectedItem", typeof(object), typeof(SelectableTreeView),
-                new UIPropertyMetadata(null, OnSelectedItemChanged));
-
-        private static void OnSelectedItemChanged(DependencyObject sender, DependencyPropertyChangedEventArgs e)
-        {
-            var item = e.NewValue as TreeViewItem;
-            if (item != null)
-            {
-                item.SetValue(TreeViewItem.IsSelectedProperty, true);
-            }
-        }
-
-        private void OnTreeViewSelectedItemChanged(object sender, RoutedPropertyChangedEventArgs<object> e)
-        {
-            SelectedItem = e.NewValue;
-        }
-    }
-}
